refactor: 구글 로그인 버튼 firebase 의존성 제거 및 components 디렉토리로 이동
